refactor: Update import paths and segment names
