+++ src/app/analyze/analytics-table.tsx
@@ -2,16 +2,16 @@
 
 import * as React from "react";
 import { useState } from "react";
-import { columns } from "./columns";
-import { data as initialData } from "./data";
-import { filterFields } from "./constants";
+import { columns } from "./const/columns";
+import { data as initialData } from "./const/data";
+import { filterFields } from "./const/constants";
 import { AnalyticsTableCore } from "./analytics-table-core";
-import { searchParamsParser } from "./search-params";
+import { searchParamsParser } from "./const/search-params";
 import { defaultAggregations } from "../../components/data-table/aggregations";
 import { useLocalStorage } from "@/hooks/use-local-storage";
 import { useQueryStates } from "nuqs";
 import { VisibilityState } from "@tanstack/react-table";
-import { ColumnSchema } from "./types";
+import { ColumnSchema } from "./types/types";
 import { cn } from "@/lib/utils";
 import { DataTableFilterControls } from "@/components/data-table/data-table-filter-controls";
 import { DataTableFilterCommand } from "@/components/data-table/data-table-filter-command";
@@ -19,8 +19,8 @@ import { DataTableToolbar } from "@/components/data-table/data-table-toolbar";
 import { DataTableFooterButtons } from "@/components/data-table/data-table-footer-buttons";
 import { DataTableGroupButtons } from "@/components/data-table/data-table-group-buttons";
 import { DataTablePagination } from "@/components/data-table/data-table-pagination";
-import { ColumnInfoTooltip } from "./column-info-tooltip";
-import { RowEditModal } from "./row-edit-modal";
+import { ColumnInfoTooltip } from "./components/column-info-tooltip";
+import { RowEditModal } from "./components/row-edit-modal";
 import { useRowEdit } from "./hooks/use-row-edit";
 import { useColumnTooltip } from "./hooks/use-column-tooltip";
 import { ColumnFiltersState } from "@tanstack/react-table";
