+++ src/app/analyze/components/skeleton.tsx
