+++ src/app/analyze/const/common.ts
