+++ src/app/analyze/const/search-params.ts
