+++ src/app/analyze/doc/analytics-table-core.md
