+++ src/app/analyze/types/types.ts
