+++ src/app/analyze/util/formatters.ts
